feature: Add setResizePane method to splitUtils class
This commit introduces the setResizePane method in the splitUtils class. This method allows users to resize the pane size, which is restricted to a range between 0 and 100 (excluding 0 and 100). The size is specified in percentage format and will be converted to pixels or remain in percentage format based on the current pane size format.